boundary repairs after upstream #3242-#3247: upstream-gospel fn_80020AEC/Toy_80310324/mnName_8023A59C, _Toy_8030FE48 prototype, clang-tidy fixes (ifstatus, ifstock, mninfo, ftCo_0A01, gmtou_2), drop unused toy helpers

+++ src/melee/gm/gmtou_2.c
@@ -943,8 +943,8 @@ void gm_8019DF8C_OnFrame(void)
                     if (color_status != 2 && color_status != 1) {
                         if (buttons & PAD_BUTTON_X) {
                             /* Down: increment color */
-                            if ((s32) tmd->x4B8[i].x3 <
-                                (s32) ((u8) gm_80169238((u8) fn_8018F6FC(
+                            if (tmd->x4B8[i].x3 <
+                                (s32) (gm_80169238((u8) fn_8018F6FC(
                                            (CSSIconHud) tmd->x4B8[i].x1)) -
                                        1))
                             {
